animation b/w search and result
