+++ pubspec.yaml
@@ -26,6 +26,7 @@ dependencies:
   flutter_rating_bar: ^3.2.0+1
   cupertino_icons: ^0.1.3
   http: ^0.12.2
+  lottie: ^0.7.0+1
   flutter_svg: ^0.17.4 #it helps us to use SVG in our app
   google_fonts: ^1.1.0 #this allows you to easily use any fonts from fonts.google.com in your Flutter app
   flutter_launcher_icons:
@@ -59,6 +60,8 @@ flutter:
     - assets/icon.png
     - assets/images/
     - assets/icons/
+    - assets/loading.json
+    - assets/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
